Added possible cmd parameter
A user can now check whether a display setting
be used or not. It validates the configuration using
CCD API to check whether the setting would work.

Also added some further checks for things like adapter
changes to stop the apply attempt if the adapter has
changed.

And updated the help text as well to better ecplain what
this little prgram does.