Added final GDI functionality
Now GDI devicecontext functionality is complete. WinLibrary can enumerate and apply settings through the GDI, which enables us to apply refresh rate changes and BPC and rotation settings. This seems to be required as the CCD path application doesnt seem to reliably set refresh rate or BPC settings in some hardware. This should fix that issue.