Updated so stores all adapters on the machine for a profile
This has been done to ensure that the profile will still work even if the profile is a cloned profile. There were some circumstances where if an adapter had one display, and that display was a clone of another display, then the list of adapters we store wouldn't include the adapter related to the cloned display. This would mean that the profile wouldn't be patched properly when it was loaded, and would fail. This is now hopefully fixed with this change.

Also updated to v1.8.1